Script for node build
